rename packet parser headers. remove unused includes.
